# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Reddit clone API",
    "description": "API for the reddit clone app",
    "version": "1.0.0"
  },
  "host": "reddit-clone-api-production.up.railway.app/",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/posts/": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get all posts",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Post a new post",
        "description": "",
        "parameters": [
          {
            "name": "title,   body,   userId,   tagIds",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "body": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                },
                "tagIds": {
                  "example": "any"
                }
              }
            },
            "description": "The post data"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get a specific post by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Post id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "Delete a specific post by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Post id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts/{id}/comments": {
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Post a new comment",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The id of the post to comment on"
          },
          {
            "name": "body, userId",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "body": {
                  "example": "any"
                }
              }
            },
            "description": "The comment data"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get all comments for a specific post by post id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Post id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts/{postid}/comments/{id}": {
      "delete": {
        "tags": [
          "Comments"
        ],
        "summary": "Delete a comment",
        "description": "",
        "parameters": [
          {
            "name": "postid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The id of the post containing the comment"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The id of the comment to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts/{id}/votes": {
      "post": {
        "tags": [
          "Votes"
        ],
        "summary": "Post a vote",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The id of the post to vote on"
          },
          {
            "name": "value",
            "in": "body",
            "description": "The vote value, 1 for upvote & -1 for downvote",
            "required": true,
            "format": "int64",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "userId",
            "in": "body",
            "description": "The id of the voting user",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts/{postId}/votes/": {
      "delete": {
        "tags": [
          "Votes"
        ],
        "summary": "Post a vote",
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The id of the post containing the vote"
          },
          {
            "name": "value",
            "in": "body",
            "description": "The vote value, 1 for upvote & -1 for downvote",
            "required": true,
            "format": "int64",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "userId",
            "in": "body",
            "description": "The id of the voting user",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user",
        "description": "",
        "parameters": [
          {
            "name": "firstName, lastName, email",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            },
            "description": "User data"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a specific user by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tags/": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "Get all tags",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Post a new tag",
        "description": "",
        "parameters": [
          {
            "name": "title",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                }
              }
            },
            "description": "The name of the tag"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tags/{id}": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "Get a specific tag by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Tag id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tags/{tagId}/{postId}/link": {
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Link a tag to a post",
        "description": "",
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tagID, postId",
            "in": "path",
            "description": "The ids of both the post and the tag to be linked",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}